requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
image==1.5.32
Pillow==7.2.0
PyJWT==1.7.1
pytz==2020.1
six==1.15.0
sqlparse==0.4.1
uuid==1.30
whitenoise==5.2.0
